Improve speed and use latest PostHog implementation
